:sparkles: Insert method is Database class
